dftg

+++ src/discord/Discord.js
@@ -169,26 +169,45 @@ module.exports = class Discord {
 
          let commandPromise = this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands().get();
          commandPromise.then(list => {
-            list.forEach(element => {
-               let commandId = element.id;
-               this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands(commandId).delete().then( () => {
-                  this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands.post({
-                     data: {
-                        name: "r",
-                        description: "Direct access to the top Public Lists",
-                        options: [
-                           {
-                               name: "Name",
-                               description: "The name of the list",
-                               type: 3,
-                               required: true,
-                               choices: choicesArray
-                           }
-                        ]
-                     }
+            //delete if needed
+            if(count(list) > 0){
+               list.forEach(element => {
+                  let commandId = element.id;
+                  this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands(commandId).delete().then( () => {
+                     this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands.post({
+                        data: {
+                           name: "r",
+                           description: "Direct access to the top Public Lists",
+                           options: [
+                              {
+                                  name: "Name",
+                                  description: "The name of the list",
+                                  type: 3,
+                                  required: true,
+                                  choices: choicesArray
+                              }
+                           ]
+                        }
+                     });
                   });
                });
-            });
+            }else{
+               this.client.api.applications(this.client.user.id).guilds(fetchedGuild.id).commands.post({
+                  data: {
+                     name: "r",
+                     description: "Direct access to the top Public Lists",
+                     options: [
+                        {
+                            name: "Name",
+                            description: "The name of the list",
+                            type: 3,
+                            required: true,
+                            choices: choicesArray
+                        }
+                     ]
+                  }
+               });
+            }
          });
       });
    }
